# resources/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Arkon API",
        "version": "1.0.0",
        "description": "Arkon's public developer API. Guide: docs/api.md. Anonymous requests read published content; a personal access token (Authorization: Bearer) adds the scopes it was given, never more than its user's role allows. Errors: {\"error\": {\"code\", \"message\", \"fields\"?}}."
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "The site the API is reached at"
        }
    ],
    "security": [
        [],
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/forms/{id}/submissions": {
            "post": {
                "operationId": "submitForm",
                "tags": [
                    "Forms"
                ],
                "summary": "Send a submission (validated, stored encrypted, rate limited)",
                "security": [
                    [],
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "The id.",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": false,
                        "description": "Makes a retry safe: the same key with the same body returns the first result; with another body it is a 409.",
                        "schema": {
                            "type": "string",
                            "pattern": "^[A-Za-z0-9_-]{16,100}$"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmissionWrite"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Received",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SubmissionResult"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Replayed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "data"
                                    ],
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/SubmissionResult"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/Error"
                    },
                    "409": {
                        "$ref": "#/components/responses/Error"
                    },
                    "410": {
                        "$ref": "#/components/responses/Error"
                    },
                    "422": {
                        "$ref": "#/components/responses/Error"
                    },
                    "429": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "A personal access token (arkon_pat_…). Scopes: read:pages, write:pages, read:posts, write:posts, write:categories, write:tags, read:media, write:media, read:forms, read:form_entries, read:users."
            }
        },
        "responses": {
            "Error": {
                "description": "An error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "object",
                        "required": [
                            "code",
                            "message"
                        ],
                        "properties": {
                            "code": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "fields": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
...
            "SubmissionWrite": {
                "type": "object",
                "required": [
                    "fields"
                ],
                "properties": {
                    "fields": {
                        "type": "object",
                        "description": "Values by field id."
                    }
                }
            },
            "SubmissionResult": {
                "type": "object",
                "required": [
                    "form",
                    "form_version",
                    "received",
                    "replayed",
                    "confirmation"
                ],
                "properties": {
                    "form": {
                        "type": "string"
                    },
                    "form_version": {
                        "type": "integer"
                    },
                    "received": {
                        "type": "boolean"
                    },
                    "replayed": {
                        "type": "boolean"
                    },
                    "confirmation": {
                        "type": "object",
                        "required": [
                            "type"
                        ],
                        "properties": {
                            "type": {
                                "type": "string",
                                "enum": [
                                    "message",
                                    "redirect"
                                ]
                            },
                            "message": {
                                "type": "string"
                            },
                            "url": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
...
        }
    }
}